requirements.txt
# Base packages for inference
inference==0.14.1

# Specify compatible versions for conflicting dependencies
botocore==1.31.64
pydantic<2.0
colorama>=0.4.5

# Additional dependencies commonly used in data science projects
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0

# Flask, in case you need it for web applications
Flask>=2.0.0

# Environment variable management
python-dotenv>=0.19.0
 